fix(memory): address PR review blockers and majors for tool-scoped memory
- Gate prefetch behind `learning.enabled` so users who opt out of learning
  don't have stored rules pinned into the system prompt
- Move prefetch after tool-set resolution and pass actual agent tool names
  to `rules_for_prompt` so unrelated tool namespaces are never scanned
- Tighten "stop" edict detection: only treat "stop " as an imperative when
  it appears at a sentence boundary; remove `contains(" stop ")` from the
  per-line check to prevent false-positive captures from phrases like
  "I want to stop working on this"
- Remove body content from repeated-failure debug log to avoid PII leaking
  into log files (log body_len only, matching the edict capture path)
- Remove duplicate `is_pinned` predicate — `is_eager` already covers both
  Critical and High; update the doc-comment to explain compression semantics
- Filter `__unscoped__` sentinel from `list_tool_names` so unscoped edicts
  captured before any tool call are not injected into future prompt filters
- Add `log::debug!` entry to all six tool-memory RPC handlers per project
  convention (stable grep-able `[tool-memory]` prefix)
- Remove raw issue reference from `tool_memory_capture` registration log
- Exercise `PromptSection::build()` in `section_renders_via_prompt_section_trait`
  test — the previous version only called `is_empty()`, leaving the trait
  contract uncovered
- Extract shared `MockMemory` to `tool_memory/test_helpers.rs`; use it in
  both `store_tests.rs` and `capture.rs` to eliminate silent drift risk
